backend-microservices/auth-service/requirements.txt
Django==5.2.12
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
django-cors-headers==4.6.0
redis==5.2.1
python-decouple==3.8
requests==2.32.3
mysqlclient==2.2.4
pytest==8.3.0
pytest-django==4.9.0
gunicorn==22.0.0
